refactor(P4-3/P4-4): フォルダ/IF整理とmilestoneのハウススタイル統一
- P4-4: domains/work-records フォルダを単数形 domains/work-record へ改名（importer更新）
- P4-3: excel-sync の live IF 2件（IExcelWbsRepository/ISyncLogRepository）を
  applications/wbs-sync/ へ移動し excel-sync フォルダを解消（importer更新）
- P4-4: Milestone をハウススタイルへ統一
  - public constructor → private constructor
  - rebuild → createFromDb（DB再構築の repository を createFromDb 呼び出しに変更）

本番ビルド成功、ソース型チェック・ユニットテスト 1292 件 green。
