+++ src/__integration_tests__/applications/wbs-sync/wbs-sync-application.integration.test.ts
@@ -2,9 +2,9 @@ import { container } from '@/lib/inversify.config';
 import { SYMBOL } from '@/types/symbol';
 import { WbsSyncApplicationService } from '@/applications/wbs-sync/wbs-sync-application.service';
 import { ExcelWbs } from '@/domains/sync/ExcelWbs';
-import type { IExcelWbsRepository } from '@/applications/excel-sync/IExcelWbsRepository';
+import type { IExcelWbsRepository } from '@/applications/wbs-sync/IExcelWbsRepository';
 import type { IWbsRepository } from '@/applications/wbs/iwbs-repository';
-import type { ISyncLogRepository } from '@/applications/excel-sync/ISyncLogRepository';
+import type { ISyncLogRepository } from '@/applications/wbs-sync/ISyncLogRepository';
 import type { IPhaseRepository } from '@/applications/task/iphase-repository';
 import type { IWbsAssigneeRepository } from '@/applications/wbs/iwbs-assignee-repository';
 import type { ITaskRepository } from '@/applications/task/itask-repository';
