+++ src/__tests__/applications/task-scheduling/scheduling-task-mapper.test.ts
@@ -6,7 +6,7 @@ import { Period } from "@/domains/task/period";
 import { PeriodType } from "@/domains/task/value-object/period-type";
 import { ManHour } from "@/domains/task/man-hour";
 import { ManHourType } from "@/domains/task/value-object/man-hour-type";
-import { WorkRecord } from "@/domains/work-records/work-record";
+import { WorkRecord } from "@/domains/work-record/work-record";
 import { Assignee } from "@/domains/task/assignee";
 import { Phase } from "@/domains/phase/phase";
 import { PhaseCode } from "@/domains/phase/phase-code";
