+++ src/applications/wbs-sync/wbs-sync-application.service.ts
@@ -2,8 +2,8 @@ import { injectable, inject } from 'inversify';
 import { ExcelWbs, SyncResult, SyncError, SyncErrorType, ValidationError } from '@/domains/sync/ExcelWbs';
 import { Task } from '@/domains/task/task';
 import { WbsDataMapper } from '@/domains/sync/WbsDataMapper';
-import type { IExcelWbsRepository } from '@/applications/excel-sync/IExcelWbsRepository';
-import type { ISyncLogRepository } from '@/applications/excel-sync/ISyncLogRepository';
+import type { IExcelWbsRepository } from '@/applications/wbs-sync/IExcelWbsRepository';
+import type { ISyncLogRepository } from '@/applications/wbs-sync/ISyncLogRepository';
 import { SYMBOL } from '@/types/symbol';
 import type { IPhaseRepository } from '@/applications/task/iphase-repository';
 import { WbsAssignee } from '@/domains/wbs/wbs-assignee';
