+++ src/infrastructures/task-repository.ts
@@ -9,7 +9,7 @@ import { PeriodType } from "@/domains/task/value-object/period-type";
 import { TaskStatus } from "@/domains/task/value-object/task-status";
 import { Task } from "@/domains/task/task";
 import { TaskNo } from "@/domains/task/value-object/task-id";
-import { WorkRecord } from "@/domains/work-records/work-record";
+import { WorkRecord } from "@/domains/work-record/work-record";
 import prisma from "@/lib/prisma/prisma";
 import { $Enums } from "@prisma/client";
 import { injectable } from "inversify";
